Improved stories & changed directory
