+++ frontend/TravelExpensePro/package.json
@@ -23,9 +23,13 @@
     "@storybook/addon-interactions": "^8.4.2",
     "@storybook/addon-onboarding": "^8.4.2",
     "@storybook/blocks": "^8.4.2",
+    "@storybook/jest": "^0.2.2",
     "@storybook/react": "^8.4.2",
     "@storybook/react-vite": "^8.4.2",
     "@storybook/test": "^8.4.2",
+    "@storybook/test-runner": "^0.19.1",
+    "@storybook/testing-library": "^0.2.1",
+    "@testing-library/react": "^16.0.1",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react-swc": "^3.5.0",
@@ -35,6 +39,7 @@
     "eslint-plugin-react-refresh": "^0.4.14",
     "eslint-plugin-storybook": "^0.11.0",
     "globals": "^15.11.0",
+    "jest": "^29.7.0",
     "prop-types": "^15.8.1",
     "storybook": "^8.4.2",
     "vite": "^5.4.10"
